requirements.txt
bottle
click
numpy
matplotlib
fluent-logger